refactor

+++ app/Http/Controllers/HouseController.php
@@ -25,7 +25,9 @@ class HouseController extends Controller
         $filename = $request->file('photo')->store('houses');
 
         // TASK: Delete the old file from the storage
-        Storage::delete($house->photo);
+        if (Storage::exists($house->photo)) {
+            Storage::delete($house->photo);
+        }
 
         $house->update([
             'name' => $request->name,
